1446 : Extraction of json block before parse.
Added support for injecting custom ServiceOutputParser to AiService
